requirements.txt
django>=2.2.0
mysqlclient==1.3.13
gunicorn==19.9.0
redis==3.3.4
djangorestframework==3.10.2
django-guardian==2.0.0
django-cors-headers==3.0.2
djangorestframework-jwt==1.11.0
coreapi==2.3.3
pygments==2.5.2
markdown==3.1.1